add AppHome with First Step Form
