+++ src/components/Application/GiveAwayForm/FormStep1.js
@@ -1,9 +1,9 @@
 import React, {useState, useContext} from 'react';
-import { FormContext } from "./FormContext";
+import {FormContext} from "./FormContext";
 
 function FormStep1() {
   
-  const {formData, setFormData}= useContext(FormContext)
+  const {formData, setFormData} = useContext(FormContext)
   
   const [type, setType] = useState(formData.type)
   
@@ -21,46 +21,61 @@ function FormStep1() {
     }));
   }
   
-
-  if(formData.step !== 1) return null
+  
+  if (formData.step !== 1) return null
   
   return (
-    <>
-      <h1 className={'appPulpit__title'}>Zaznacz co chcesz oddać</h1>
-      <form className={'appPulpit__form'} onSubmit={handleSubmit}>
-        <label>
-          <input type={'radio'} value={'ubrania do ponownego użycia'}
-                 checked={type === 'ubrania do ponownego użycia'}
-                 onChange={handleChange}/>
-          Ubrania do ponownego użycia
-        </label>
-        <label>
-          <input type={'radio'} value={'ubrania do wyrzucenia'}
-                 checked={type === 'ubrania do wyrzucenia'}
-                 onChange={handleChange}/>
-          Ubrania do wyrzucenia
-        </label>
-        <label>
-          <input type={'radio'} value={'zabawki'}
-                 checked={type === 'zabawki'}
-                 onChange={handleChange}/>
-          Zabawki
-        </label>
-        <label>
-          <input type={'radio'} value={'książki'}
-                 checked={type === 'książki'}
-                 onChange={handleChange}/>
-          Książki
-        </label>
-        <label>
-          <input type={'radio'} value={'inne'}
-                 checked={type === 'inne'}
-                 onChange={handleChange}/>
-          Inne
-        </label>
-      </form>
-      <button onClick={handleSubmit} type={"submit"} className={'btn btn-small'}>Dalej</button>
-    </>
+    <section className={'formStep'}>
+      <div className={'warning'}>
+        <div className={'container'}>
+          <h2 className={'warning__title'}>Ważne!</h2>
+          <p className={'warning__text'}>Uzupełnij szczegóły dotyczące Twoich rzeczy. Dzięki temu będziemy wiedzieć komu
+            najlepiej je przekazać.</p>
+        </div>
+      </div>
+      <div className={'container'}>
+        <p className={'formStep__counter'}>Krok {formData.step}/4</p>
+        <h1 className={'formStep__title'}>Zaznacz co chcesz oddać</h1>
+        <form className={'formStep__form'} onSubmit={handleSubmit}>
+          <label className={'formStep__label'}>
+            <input type={'radio'} value={'ubrania do ponownego użycia'}
+                   checked={type === 'ubrania do ponownego użycia'}
+                   onChange={handleChange}
+                   className={'formStep__input'}/>
+            Ubrania do ponownego użycia
+          </label>
+          <label className={'formStep__label'}>
+            <input type={'radio'} value={'ubrania do wyrzucenia'}
+                   checked={type === 'ubrania do wyrzucenia'}
+                   onChange={handleChange}
+                   className={'formStep__input'}/>
+            Ubrania do wyrzucenia
+          </label>
+          <label className={'formStep__label'}>
+            <input type={'radio'} value={'zabawki'}
+                   checked={type === 'zabawki'}
+                   onChange={handleChange}
+                   className={'formStep__input'}/>
+            Zabawki
+          </label>
+          <label className={'formStep__label'}>
+            <input type={'radio'} value={'książki'}
+                   checked={type === 'książki'}
+                   onChange={handleChange}
+                   className={'formStep__input'}/>
+            Książki
+          </label>
+          <label className={'formStep__label'}>
+            <input type={'radio'} value={'inne'}
+                   checked={type === 'inne'}
+                   onChange={handleChange}
+                   className={'formStep__input'}/>
+            Inne
+          </label>
+        </form>
+        <button onClick={handleSubmit} type={"submit"} className={'btn btn-small'}>Dalej</button>
+      </div>
+    </section>
   );
 }
