+++ src/scss/Application/_appForm.scss
@@ -1,0 +1,21 @@
+.appForm {
+  background-image: url("../../assets/Background-Form.jpg");
+  background-position: center;
+  //background-position: -30vw;
+  background-repeat: no-repeat;
+  background-size: cover;
+  height: 100vh;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+
+
+  &__title {
+
+  }
+
+  &__form {
+
+  }
+}
