--- src/scss/Application/_appPulpit.scss
@@ -1,19 +1,0 @@
-.appPulpit {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 2.0rem;
-
-  &__title {
-    font-size: 3.0rem;
-    margin: 50px 0;
-  }
-
-  &__form {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    width: 300px;
-    margin: 50px 0;
-  }
-}
