+++ src/scss/Application/_formStep.scss
@@ -1,0 +1,68 @@
+.formStep {
+  display: flex;
+  flex-direction: column;
+  //align-items: flex-start;
+  position: relative;
+  width: 100%;
+  height: 100%;
+  padding: 100px 0 0;
+
+
+  &__counter {
+    font-size: 1.5rem;
+    font-weight: 300;
+    margin: 20px 0 40px;
+  }
+
+  &__title {
+    font-size: 2.5rem;
+    font-weight: 600;
+    margin: 0 0 30px;
+  }
+
+  &__form {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+margin-bottom: 100px;
+  }
+
+  &__label {
+    margin: 10px 0;
+    font-size: 1.8rem;
+    font-weight: 300;
+  }
+
+  &__input {
+width: 20px;
+  }
+}
+
+.warning {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: center;
+  width: 100%;
+  height: 100px;
+  position: absolute;
+  top: 0;
+  background-color: $color-yellow;
+
+.container{
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: center;
+}
+  &__title {
+    font-size: 2.0rem;
+    font-weight: 600;
+    margin: 0 0 5px;
+  }
+
+  &__text {
+    font-size: 1.5rem;
+
+  }
+}
