+++ src/scss/main.scss
@@ -19,5 +19,6 @@
 @import "Application/login";
 @import "Application/signup";
 @import "Application/logout";
-@import "Application/appPulpit";
+@import "Application/appForm";
 @import "Application/appMain";
+@import "Application/formStep";
